+++ evadb/storage/image_storage_engine.py
@@ -27,7 +27,7 @@ class ImageStorageEngine(AbstractMediaStorageEngine):
         super().__init__(db)
 
     def read(self, table: TableCatalogEntry) -> Iterator[Batch]:
-        for image_files in self._rdb_handler.read(self._get_metadata_table(table), 12):
+        for image_files in self._rdb_handler.read(self._get_metadata_table(table)):
             for _, (row_id, file_name) in image_files.iterrows():
                 system_file_name = self._xform_file_url_to_file_name(file_name)
                 image_file = Path(table.file_url) / system_file_name